displaying tables on certain conditions

+++ homepage.php
@@ -12,7 +12,9 @@
 
 	<?php
 		session_start();
+		$user_set = false;
 		if (isset($_SESSION['user'])) {
+			$user_set = true;
 			require 'database.php';
 			$user_id=$_SESSION['user'];
 			$stmt = $mysqli->prepare("select user from users where user_id=\"".$user_id."\"");
@@ -56,8 +58,12 @@
 		<tr>
 			<th>Title</th>
 			<th>Author</th>
-			<th>Edit</th>
-			<th>Delete</th>
+			<?php
+			if ($user_set) {
+				echo "<th>Edit</th>";
+				echo "<th>Delete</th>";
+			}
+			?>
 		</tr>  
 		</thead>
 		<tbody>
@@ -78,8 +84,10 @@ while($stmt->fetch()){
 	echo "<tr>\n";
 	printf("\t<td> <a href=\"story.php?id=$story_id\">%s</a></td>\n", htmlspecialchars($title));
 	printf("\t<td>%s</td>\n", htmlspecialchars($author));
-	echo "\t<td> <a href=\"edit.php?id=$story_id\" class=\"button\">Edit</a></td>\n";
-	echo "\t<td> <a href=\"delete.php?id=$story_id\" class=\"button\">Delete</a></td>\n";
+	if ($user_set) {
+		echo "\t<td> <a href=\"edit.php?id=$story_id\" class=\"button\">Edit</a></td>\n";
+		echo "\t<td> <a href=\"delete.php?id=$story_id\" class=\"button\">Delete</a></td>\n";
+	}
 	echo "</tr>\n";
 }
 echo "</tbody>\n";
